+++ src/main/kotlin/org/bibletranslationtools/fetcher/application/SourceAudioRepository.kt
@@ -4,9 +4,9 @@ import java.nio.file.Files
 import java.nio.file.Paths
 import kotlin.streams.toList
 import org.bibletranslationtools.fetcher.data.Language
-import org.bibletranslationtools.fetcher.domain.LanguageRepository
+import org.bibletranslationtools.fetcher.domain.LanguageCatalog
 
-class SourceAudioLanguages(private val languageRepository: LanguageRepository) {
+class SourceAudioLanguages(private val languageRepository: LanguageCatalog) {
 
     private val sourceAudioRoot = "/SourceAudio"
